Move `PathInfo` into its own module, rm `describeObject()`, rearrange `utils.js` a little.
